requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
SQLAlchemy==1.4.31
psycopg2-binary==2.9.1
PyJWT==2.3.0
passlib==1.7.5
